Remove unused ResourceForm component and move Resource definitions and functions to src/models/
